client finish v!

+++ src/pages/cart.tsx
@@ -8,7 +8,7 @@ import { useNavigate, useParams } from 'react-router-dom'
 
 const Cart = () => {
   const isFirstRender = useFirstRender()
-  const { tg } = useTelegram()
+  const { tg, user } = useTelegram()
   const navigate = useNavigate()
   const { locale } = useParams()
 
@@ -34,7 +34,7 @@ const Cart = () => {
     try {
       if (isFirstRender) setIsLoading(true)
 
-      await CartApi.getAll().then((data) => {
+      await CartApi.getAll(user.id).then((data) => {
         setCartItem(data.baskets)
         setTotalPrice(data.total_price)
         setDeliverPrice(data.deleviry_price)
@@ -53,7 +53,7 @@ const Cart = () => {
   const handleDelete = async (id: number) => {
     try {
       setDeleteIsloading(id)
-      await CartApi.delete(id)
+      await CartApi.delete({ id, user_id: user.id })
       getBasket()
     } catch (error) {
       console.log(error)
@@ -65,7 +65,7 @@ const Cart = () => {
   const handleAddCount = async (id: number) => {
     try {
       setPlusLoading(id)
-      await CartApi.changeCount({ id, body: { count: 1 } })
+      await CartApi.changeCount({ id, body: { count: 1 }, user_id: user.id })
       getBasket()
     } catch (error) {
       console.log(error)
@@ -78,7 +78,7 @@ const Cart = () => {
     try {
       if (count > 1) {
         setMinusLoading(id)
-        await CartApi.changeCount({ id, body: { count: -1 } })
+        await CartApi.changeCount({ id, body: { count: -1 }, user_id: user.id })
 
         getBasket()
       }
